AI/prison_health_api/requirements.txt
flask
flask-sqlalchemy
flask-cors
ultralytics
opencv-python
langchain
langchain-google-genai
langchain-community
chromadb
pypdf
python-dotenv
langchain-text-splitters